# internal/interfaces/http/swagger/swagger.yaml
definitions:
  invoice.Invoice:
    properties:
      amount:
        type: number
      id:
        type: string
      status:
        $ref: '#/definitions/invoice.Status'
      user_id:
        type: string
    type: object
  invoice.Status:
    enum:
    - pending
    - paid
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusPaid
  response.ErrorResponse:
    properties:
      error:
        example: invalid request
        type: string
    type: object
info:
  contact: {}
paths:
  /invoices:
    get:
      consumes:
      - application/json
      description: show all invoice
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: All Invoice
      tags:
      - Invoice
  /invoices/{id}:
    get:
      description: Retrieve a single invoice by its ID
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/invoice.Invoice'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get invoice by ID
      tags:
      - Invoice
  /users:
    post:
      consumes:
      - application/json
      description: Creates a new user with ID, name, email
      parameters:
      - description: User Payload
        in: body
        name: user
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - User
  /users/get:
    get:
      description: Fetch a user by query parameter id
      parameters:
      - description: User ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by ID
      tags:
      - Get Users
swagger: "2.0"